Add a way to ignore failure when boolean do not exist
Ansible is used by various groups as a installer, and to be multi
platform, they sometime need to set selinux boolean
who may or may not be present in the current policy.
Testing them is cumbersome, so ignoring that specific error
permit to reduce code.

See https://github.com/ansible/ansible-modules-core/pull/4570#issuecomment-242858439